Раздел 8

+++ src/service/data-service/comment.js
@@ -5,25 +5,48 @@ class CommentService {
     this._Offer = sequelize.models.Offer;
     this._Comment = sequelize.models.Comment;
   }
-  create(offerId, comment) {
-    return this._Comment.create({
+
+  async create(offerId, comment) {
+    return await this._Comment.create({
       offerId,
       ...comment
     });
   }
 
-  async drop(id) {
+  async drop(userId, offerId, commentId) {
+    const offerByUser = await this._Offer.findOne({
+      where: {
+        id: offerId,
+        userId
+      }
+    });
+
+    if (!offerByUser) {
+      return !!offerByUser;
+    }
+
     const deletedRows = await this._Comment.destroy({
-      where: {id}
+      where: {
+        id: commentId
+      }
     });
 
     return !!deletedRows;
   }
 
-  findAll(offerId) {
-    return this._Comment.findAll({
+  async findAll(offerId) {
+    return await this._Comment.findAll({
       where: {offerId},
       raw: true
+    });
+  }
+
+  async findOne(id, offerId) {
+    return await this._Comment.findOne({
+      where: {
+        id,
+        offerId
+      }
     });
   }
